--- eslint.config.mjs
@@ -1,21 +1,0 @@
-import globals from "globals";
-import pluginJs from "@eslint/js";
-
-export default [
-  {
-    languageOptions: {
-      globals: {
-        Morebits: "readonly",
-        mw: "readonly",
-        ...globals.browser,
-        Window: "writable"
-      }
-    }
-  },
-  pluginJs.configs.recommended,
-
-  {
-    rules: {}
-  }
-
-];
