+++ src/modules/reports.ts
@@ -267,7 +267,7 @@ function submitMessage(e: Event) {
     }
 }
 
-function createFormWindow(reportedUserFromDOM: string | null) {
+function createFormWindow(reportedUserFromDOM: string | null): void {
 
     // Something about the addPortletLink feature doesn't work well so this condition is unfortunately needed
     if (typeof reportedUserFromDOM == 'string') {
